Removed the transmission-model kappa-update line from the change-health-system intervention.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@627 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/Survey.h
@@ -129,7 +129,7 @@ class SurveyAgeGroup {
     
     /// Get the total number of age categories (inc. one for indivs. not in any
     /// category given in XML).
-    static inline int getNumGroups () {
+    static inline size_t getNumGroups () {
       return _upperbound.size();
     }
